Initial attempt at splitting Resources into separate files
